requirements.txt
streamlit
torch
torchvision
Pillow
numpy
timm
scikit-learn
matplotlib
opencv-python
clip
transformers